Trick ts-node to correctly resolve files in pack-a
How ts-node choses to resolve dependencies in submodules seems to
be tricker than I though. But by adding an empty nop.ts to the root
of the pack-a workspace, ts-node will correctly assume that the root
of the workspace is that folder, and thus pack-a will also work in
dev mode.